requirements.txt
yt-dlp
openai-whisper
pathvalidate
tqdm